feat: notify user if username/password incorrect
If the username and/or password is incorrect, the user will now be notified
and be given the chance to enter them again. Before the program would
continue normally and then cause an error when trying to analyze the call log
since it could not analyze the error thrown by the Cudatel Communications
Server.